requirements.txt
torchvision
torchaudio
audiocraft==1.3.0
laion-clap==1.1.4
huggingface_hub==0.21.4 
bittensor==7.4.0
GPUtil==1.4.0
encodec==0.1.1
funcy==2.0
transformers==4.37.2
audiolm-pytorch==2.0.7
speechbrain==0.5.16
datasets==2.16.1
librosa==0.10.0.post2
jiwer==3.0.3
matplotlib==3.8.2
uvicorn==0.22.0 
pyngrok==7.1.2
wandb==0.16.2
SQLAlchemy==1.4.51
passlib==1.7.4 
psycopg2-binary==2.9.9
python-jose==3.3.0
inflect
cached_path 
gruut[de,es,fr]==2.2.3
pykakasi==2.2.1 
fugashi==1.3.0 
g2p_en==2.1.0 
anyascii==0.3.2 
jamo==0.4.1
cn2an==0.5.22 
pypinyin==0.50.0 
mecab-python3==1.0.5 
num2words==0.5.12 
unidic_lite==1.0.8 
unidic==1.1.0
jieba==0.42.1
tensorboardX
pyOpenSSL